alterando migrations e seeds

+++ src/database/migrations/20220718041045-create-products.js
@@ -23,7 +23,7 @@ module.exports = {
                 allowNull: false,
             },
             price: {
-                type: Sequelize.INTEGER,
+                type: Sequelize.DECIMAL(20, 2),
                 allowNull: false,
             },
             stock: {
